actualizado

+++ src/main/java/com/upc/appsaludai3/security/repositories/RoleRepository.java
@@ -1,8 +1,7 @@
 package com.upc.appsaludai3.security.repositories;
 
-
 import com.upc.appsaludai3.security.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 }
