+++ src/main/java/com/upc/appsaludai3/security/repositories/UserRepository.java
@@ -2,16 +2,17 @@ package com.upc.appsaludai3.security.repositories;
 
 
 
+
 import com.upc.appsaludai3.security.entities.User;
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 
     @Transactional
